added hierarchy in devices module
