+++ src/qnngds/__init__.py
@@ -1,3 +1,11 @@
-def main():
-    """Entry point for the application script."""
-    print("Call your main application code here")
+from . import devices, circuits, design, geometries, tests, utilities
+
+
+def help():
+    """ Provides links to documentation """
+
+    print("Need help? Check qnngds documentation: https://qnngds.readthedocs.io/en/latest/ \n")
+    print(" - Tutorials: https://qnngds.readthedocs.io/en/latest/tutorials.html \n")
+    print(" - API: https://qnngds.readthedocs.io/en/latest/api.html \n\n")
+
+    print("You are a contributor? Check this documentation: https://qnngds.readthedocs.io/projects/qnngds-dev/en/latest/")
