+++ src/qnngds/devices/__init__.py
@@ -1,0 +1,4 @@
+""" Devices module contains variations of common devices used in the circuits
+designed by QNN group. """
+
+from . import htron, nanowire, ntron, resistor, snspd
