+++ src/qnngds/devices/nanowire.py
@@ -1,0 +1,35 @@
+from phidl import Device
+
+import phidl.geometry as pg
+from typing import Tuple, Optional
+
+
+def spot(
+    channel_w: float = 0.1, source_w: float = 0.3, layer: int = 0, num_pts: int = 100
+) -> Device:
+    """Creates a single wire, with the same appearance as an NTRON but without
+    the gate.
+
+    Args:
+        channel_w (int or float): The width of the channel (at the hot-spot location).
+        source_w (int or float): The width of the nanowire's "source".
+        layer (int): The layer where to put the device.
+        num_pts (int): The number of points comprising the optimal_steps geometries.
+
+    Returns:
+        Device: A device containing 2 optimal steps joined at their channel_w end.
+    """
+
+    NANOWIRE = Device(f"NANOWIRE {channel_w} ")
+    wire = pg.optimal_step(channel_w, source_w, symmetric=True, num_pts=num_pts)
+    source = NANOWIRE << wire
+    gnd = NANOWIRE << wire
+    source.connect(source.ports[1], gnd.ports[1])
+
+    NANOWIRE.flatten(single_layer=layer)
+    NANOWIRE.add_port(name=1, port=source.ports[2])
+    NANOWIRE.add_port(name=2, port=gnd.ports[2])
+    NANOWIRE.rotate(-90)
+    NANOWIRE.move(NANOWIRE.center, (0, 0))
+
+    return NANOWIRE
